bug fixeeees

+++ src/objects/ProgressBestRun.hpp
@@ -4,8 +4,8 @@
 
 struct ProgressBestRun
 {
-    uint32_t m_fromPercent = 0;
-    uint32_t m_toPercent = 0;
+    float m_fromPercent = 0;
+    float m_toPercent = 0;
 
     void reset() {
         m_fromPercent = m_toPercent = 0;
